Center position of loading message
I think this is better default, and indirectly hint how this process can be customized. 
I also would like to submit this change to templates repo, because without that startup experience degraded. Given that current template is somewhat fully functional, I believe that small tidbit would be helpful for developers